[ poc ] port CSE example to Snoclist + fix bug
The bug is a bit subtle: when deciding whether to abstract a
subterm, I would check the scopes for equality and then the
terms. But in co-de Bruijn that's not a good idea:

Both

[<Bool] ---[<100]--> [< Bool, Nat, Bool]
[<Bool] ---[<001]---^

restrict [< Bool, Nat, Bool] to the same support [<Bool]
and yet the term (V Here) will have a different meaning
depending on which thinning was picked: either (V 0) or (V 2).

So it's not enough to check that the supports are the same, we
need to check that the thinnings are.