[ cleanup ] We can scrap more boilerplate of the boilerplate.
Given that reduction *is* the reflexive transitive closure for
reduction, let us use a generic definition.  Thanks to Frex for this.
We have included the Frexy code until their generic code is merged
into Idris2's libs.

We also not that `Finished` is a glorified `Maybe`. So let's use that.

Thanks @gallais!