[ more generic ast ] ASTs *can* contain other ASTs
The generic AST library has been kinded to allow for a simple stratification of nodes based on different _kinds_ of AST.

Velo has been updated to host the new adjustments.

If we were to delay the construction of types from the parser to the elaborator then we can include types in the AST itself.